+++ package.json
@@ -9,16 +9,20 @@
   },
   "dependencies": {
     "@3id/connect": "0.2.2",
+    "@glazed/datamodel": "^0.2.0",
+    "@glazed/did-datastore": "^0.2.1",
     "@self.id/framework": "^0.1.0",
     "@walletconnect/web3-provider": "^1.6.6",
     "ethers": "^5.5.1",
     "formik": "^2.2.9",
+    "framer-motion": "^5.5.1",
     "ipfs-core": "^0.12.0",
     "ipfs-http-client": "^54.0.0",
     "next": "12.0.3",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-dropzone": "^11.4.2",
+    "react-flip-ticker": "^1.0.2",
     "react-player": "^2.9.0",
     "react-tsparticles": "^1.37.4",
     "styled-components": "^5.3.3",
