+++ pages/_app.js
@@ -1,89 +1,268 @@
-import Particles from 'react-tsparticles'
-import '../styles/globals.css'
+import Particles from 'react-tsparticles';
+import '../styles/globals.css';
+import Web3Modal from 'web3modal';
+import { providers } from 'ethers';
+import { useCallback, useEffect, useReducer } from 'react';
+import WalletConnectProvider from '@walletconnect/web3-provider';
+import styled from 'styled-components';
+
+const providerOptions = {
+  walletconnect: {
+    package: WalletConnectProvider, // required
+    options: {
+      infuraId: 'a83b0fa56b894d9daa4aaecc4fbfe712', // required
+    },
+  },
+};
+
+let web3Modal;
+if (typeof window !== 'undefined') {
+  web3Modal = new Web3Modal({
+    network: 'mainnet', // optional
+    cacheProvider: true,
+    providerOptions, // required
+  });
+}
+
+const initialState = {
+  provider: null,
+  web3Provider: null,
+  address: null,
+  chainId: null,
+};
+
+function reducer(state, action) {
+  switch (action.type) {
+    case 'SET_WEB3_PROVIDER':
+      return {
+        ...state,
+        provider: action.provider,
+        web3Provider: action.web3Provider,
+        address: action.address,
+        chainId: action.chainId,
+      };
+    case 'SET_ADDRESS':
+      return {
+        ...state,
+        address: action.address,
+      };
+    case 'SET_CHAIN_ID':
+      return {
+        ...state,
+        chainId: action.chainId,
+      };
+    case 'RESET_WEB3_PROVIDER':
+      return initialState;
+    default:
+      throw new Error();
+  }
+}
 
 
 function MyApp({ Component, pageProps }) {
-	return (
-		<>
-			<Particles
-				id='tsparticles'
-				style={{
-					position: 'relative',
-					zIndex: -1,
-				}}
-				options={{
-					fpsLimit: 60,
-					interactivity: {
-						events: {
-							onHover: {
-								enable: true,
-								mode: 'repulse',
-							},
-						},
-						modes: {
-							bubble: {
-								distance: 400,
-								duration: 2,
-								opacity: 0.8,
-								size: 40,
-							},
-							push: {
-								quantity: 4,
-							},
-							repulse: {
-								distance: 50,
-								duration: 0.4,
-							},
-						},
-					},
-					particles: {
-						color: {
-							value: '#ffc115',
-						},
-						links: {
-							color: '#D8F9FF',
-							distance: 150,
-							enable: true,
-							opacity: 0.1,
-							width: 1,
-						},
-						collisions: {
-							enable: true,
-						},
-						move: {
-							direction: 'none',
-							enable: true,
-							outMode: 'bounce',
-							random: false,
-							speed: 1,
-							straight: false,
-						},
-						number: {
-							density: {
-								enable: true,
-								value_area: 1000,
-							},
-							value: 80,
-						},
-						opacity: {
-							value: 0.5,
-						},
-						shape: {
-							type: 'triangle',
-						},
-						size: {
-							random: true,
-							value: 3,
-						},
-					},
-					detectRetina: true,
-				}}
-			/>
-			<div style={{ position: 'relative', zIndex: 2 }}>
-				<Component {...pageProps} />
-			</div>
-		</>
-	)
+  const [state, dispatch] = useReducer(reducer, initialState);
+  const { provider, web3Provider, address, chainId } = state;
+
+  const connect = useCallback(async function () {
+    // This is the initial `provider` that is returned when
+    // using web3Modal to connect. Can be MetaMask or WalletConnect.
+    const provider = await web3Modal.connect();
+
+    // We plug the initial `provider` into ethers.js and get back
+    // a Web3Provider. This will add on methods from ethers.js and
+    // event listeners such as `.on()` will be different.
+    const web3Provider = new providers.Web3Provider(provider);
+
+    const signer = web3Provider.getSigner();
+    const address = await signer.getAddress();
+
+    const network = await web3Provider.getNetwork();
+
+    dispatch({
+      type: 'SET_WEB3_PROVIDER',
+      provider,
+      web3Provider,
+      address,
+      chainId: network.chainId,
+    });
+  }, []);
+
+  const disconnect = useCallback(
+    async function () {
+      await web3Modal.clearCachedProvider();
+      if (provider?.disconnect && typeof provider.disconnect === 'function') {
+        await provider.disconnect();
+      }
+      dispatch({
+        type: 'RESET_WEB3_PROVIDER',
+      });
+    },
+    [provider]
+  );
+
+  // Auto connect to the cached provider
+  useEffect(() => {
+    if (web3Modal.cachedProvider) {
+      connect();
+    }
+  }, [connect]);
+
+  // A `provider` should come with EIP-1193 events. We'll listen for those events
+  // here so that when a user switches accounts or networks, we can update the
+  // local React state with that new information.
+  useEffect(() => {
+    if (provider?.on) {
+      const handleAccountsChanged = (accounts) => {
+        // eslint-disable-next-line no-console
+        console.log('accountsChanged', accounts);
+        dispatch({
+          type: 'SET_ADDRESS',
+          address: accounts[0],
+        });
+      };
+
+      // https://docs.ethers.io/v5/concepts/best-practices/#best-practices--network-changes
+      const handleChainChanged = (_hexChainId) => {
+        window.location.reload();
+      };
+
+      const handleDisconnect = (error) => {
+        // eslint-disable-next-line no-console
+        console.log('disconnect', error);
+        disconnect();
+      };
+
+      provider.on('accountsChanged', handleAccountsChanged);
+      provider.on('chainChanged', handleChainChanged);
+      provider.on('disconnect', handleDisconnect);
+
+      // Subscription Cleanup
+      return () => {
+        if (provider.removeListener) {
+          provider.removeListener('accountsChanged', handleAccountsChanged);
+          provider.removeListener('chainChanged', handleChainChanged);
+          provider.removeListener('disconnect', handleDisconnect);
+        }
+      };
+    }
+  }, [provider, disconnect]);
+
+  // const chainData = getChainData(chainId);
+  return (
+    <>
+      <Particles
+        id="tsparticles"
+        style={{
+          position: 'relative',
+          zIndex: -1,
+        }}
+        options={{
+          fpsLimit: 60,
+          interactivity: {
+            events: {
+              onHover: {
+                enable: true,
+                mode: 'repulse',
+              },
+            },
+            modes: {
+              bubble: {
+                distance: 400,
+                duration: 2,
+                opacity: 0.8,
+                size: 40,
+              },
+              push: {
+                quantity: 4,
+              },
+              repulse: {
+                distance: 50,
+                duration: 0.4,
+              },
+            },
+          },
+          particles: {
+            color: {
+              value: '#ffc115',
+            },
+            links: {
+              color: '#D8F9FF',
+              distance: 150,
+              enable: true,
+              opacity: 0.1,
+              width: 1,
+            },
+            collisions: {
+              enable: true,
+            },
+            move: {
+              direction: 'none',
+              enable: true,
+              outMode: 'bounce',
+              random: false,
+              speed: 1,
+              straight: false,
+            },
+            number: {
+              density: {
+                enable: true,
+                value_area: 1000,
+              },
+              value: 80,
+            },
+            opacity: {
+              value: 0.5,
+            },
+            shape: {
+              type: 'triangle',
+            },
+            size: {
+              random: true,
+              value: 3,
+            },
+          },
+          detectRetina: true,
+        }}
+      />
+      <div style={{ position: 'relative', zIndex: 2 }}>
+        <Header>
+          {web3Provider ? (
+            <ButtonFrame type="button" onClick={disconnect}>
+              <Text>{address.slice(0, 6)}...</Text>
+            </ButtonFrame>
+          ) : (
+            <ButtonFrame type="button" onClick={connect}>
+              <Text>Connect</Text>
+            </ButtonFrame>
+          )}
+        </Header>
+        <Component
+          address={address}
+          web3Provider={web3Provider}
+          {...pageProps}
+        />
+      </div>
+    </>
+  );
 }
 
-export default MyApp
+export default MyApp;
+
+const Header = styled.div`
+  display: flex;
+  padding: 2rem 0;
+  width: 100%;
+  justify-content: flex-end;
+`;
+
+const ButtonFrame = styled.button`
+  background-color: cadetblue;
+  padding: 0 1rem;
+  outline: none;
+  border: none;
+  border-radius: 15px;
+`;
+
+const Text = styled.p`
+  color: whitesmoke;
+`;
